pyproject.toml
[project]
name = "amanu"
version = "1.7.0"
description = "AI-powered amanuensis for your voice notes. Mp3 to Markdown with structure."
authors = [
    {name = "jfima", email = "efimix@gmail.com"},
]
dependencies = [
    "google-generativeai",
    "pyyaml",
    "watchdog",
    "rich",
    "questionary",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv",
    "openai>=1.0.0",
    "requests",
]
requires-python = ">=3.10"

[project.scripts]
amanu = "amanu.cli:main"

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"
